backend/requirements.txt
fastapi
uvicorn
python-multipart
joblib
numpy
scipy
scikit-learn
xgboost
librosa
openai-whisper
pydantic
python-dotenv
